Fix a handful of warnings and one code error: -    char* tracknum[20]; +    char tracknum[20];
svn path=/trunk/kdemultimedia/kscd/libwm/; revision=332005

+++ cddb.c
@@ -237,7 +237,7 @@ connect_open(void)
 		int inet_addr();
 		
 		defaddr.s_addr = inet_addr(host);
-		if (defaddr.s_addr == -1) 
+		if ((int) defaddr.s_addr == -1) 
 		{
 #ifndef NDEBUG
 			printf("unknown host: %s\n", host);
